requirements.txt
python-telegram-bot
requests
Pillow
pytz
rich
yt-dlp
instaloader
mutagen
telegram
flask
pytelegrambotapi
